Rename driver-filechannel to driver-file.
