+++ build.gradle
@@ -81,9 +81,9 @@ allprojects {
 dependencies {
 	compile project(":sbk-api")
 	compile project(":driver-concurrentq")
+	compile project(":driver-file")
 	compile project(":driver-filestream")
 	compile project(":driver-asyncfile")
-	compile project(":driver-filechannel")
 	compile project(":driver-hdfs")
 	compile project(":driver-bookkeeper")
 	compile project(":driver-rabbitmq")
