--- driver-filechannel/src/main/resources/filechannel.properties
@@ -1,19 +1,0 @@
-#Copyright (c) KMG. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Default file name
-fileName=./tmp.txt
-
-# Is append
-isAppend=false
-
-# Is recreate
-reCreate=true
-
-#Update meta data while flush
-metaUpdate=false
