+++ settings.gradle
@@ -15,9 +15,9 @@ include 'sbk-api'
 include 'driver-pulsar'
 include 'driver-kafka'
 include 'driver-pravega'
+include 'driver-file'
 include 'driver-filestream'
 include 'driver-asyncfile'
-include 'driver-filechannel'
 include 'driver-concurrentq'
 include 'driver-bookkeeper'
 include 'driver-template'
